Safari-611.1.10/LayoutTests/platform/glib/accessibility/inline-continuations-expected.txt

Inline continuations - accessibility tree linkage

This test checks that the right accessibility tree is generated in the presence of inline continuations. Each of the five numbers below should be visited only ones traversing the accessibility tree.

1
2
34
5
FAIL nonGroupDescendants(accessibilityController.focusedElement) should be 5. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
